test: the fork-only families build under clang with warnings as errors
clang 22 rejects ninety-three constructs in these units that GCC accepts, all of
them in the tests but one:

  - forty-nine of the reflection macro's hidden friend left unemitted, because
    the block it belongs to sits in an anonymous namespace and nothing in the
    unit calls the friend. The nineteen units now put their helper blocks in a
    named namespace with a using-directive after it, the form the other tests
    use.
  - twenty locals whose names boost::ut also declares in the scope the case runs
    in: nine RuntimeTest instances named "test", seven sample counts named
    "skip", four tag loop variables named "tag".
  - nineteen float values widened to double inside an expression, eighteen of
    them the real and imaginary parts handed to a std::complex<double>, now
    stated with a cast at the point of widening.
  - three shifts and complements whose operands are int against unsigned, one of
    them in p25/ImbeFrame.hpp, where imbeUnpackUVector() accumulates the
    parameter word in unsigned as the packing side beside it already does.
  - one aggregate initialized with braces where the refusal case needs only the
    settings, and one lambda capture of a constant the body reads by value.

The behavior of every case is unchanged: no assertion, no bound and no fixture
moves. The audio test is upstream's, unmodified here, and its six missing-field
initializers are left with it.

+++ blocks/measurement/test/qa_AutocorrelationRecording.cpp
@@ -406,7 +406,7 @@ const suite<"autocorrelation recording legs"> _acfRecording = [] {
             std::print("{:.1f} us: {:.4f}   ", 0.5 * static_cast<double>(lag), ungatedArm.magnitude[lag]);
         }
         std::println("");
-        const double flatSpread = *std::ranges::max_element(ungatedArm.magnitude.begin() + 1, ungatedArm.magnitude.begin() + 7) - *std::ranges::min_element(ungatedArm.magnitude.begin() + 1, ungatedArm.magnitude.begin() + 7);
+        const double flatSpread = static_cast<double>(*std::ranges::max_element(ungatedArm.magnitude.begin() + 1, ungatedArm.magnitude.begin() + 7)) - static_cast<double>(*std::ranges::min_element(ungatedArm.magnitude.begin() + 1, ungatedArm.magnitude.begin() + 7));
         std::println("    no lag between 0.5 and 3.0 us is distinguishable: the whole run spans {:.4f} on a curve that never falls to the scatter", flatSpread);
         expect(static_cast<double>(ungatedArm.magnitude[2UZ]) > 20. * ungatedArm.scatter) << "the derived null at 1 us is nowhere to be seen in an ungated window";
